Implement proper showdown mucking mechanics.
Robert's Rules of Poker:
" If everyone checks (or is all-in) on the final betting round, the player who acted first is the first to show the hand. If there is wagering on the final betting round, the last player to take aggressive action by a bet or raise is the first to show the hand. "