Solve Issue #67: Fix showdown delay
Instead of the client busy waiting for 5 seconds after a showdown, the server
will delay dealing the next hand for a specified number of seconds.

Use the `showdown` attribute in the `<delays />` node of the server config to
specify the number of seconds the server should wait after a showdown before
dealing the next hand. By default the `showdown` attribute has a value of `18`,
which is too long of delay. Setting the `showdown` attribute to a value of `5`
is suggested.